pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "salesforce-mcp-server"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    "mcp",
    "python-dotenv",
    "openai",  # or any other dependencies
    "requests>=2.31.0",
]

[tool.setuptools]
packages = ["src"]
py-modules = ["main"]

[project.scripts]
main = "main:main"